v0.7
Major changes to Fit.m with the addition of the multi-fit Lorentzian-Gaussian hybridization, as well as added capabilities for multi-frame selection, fitting, clearing, and exporting.  Exporting now gives .fig and .png along with the raw data.  May need some work on exactly how to export data, but we should be in the clear!